muxi_shop_api/requirements.txt
asgiref==3.10.0
axios==0.4.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
Django==4.2.25
django-cors-headers==4.9.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
djangorestframework==3.16.1
Flask==3.1.2
flask-cors==6.0.1
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mysqlclient==2.2.7
pillow==11.3.0
pycryptodome==3.23.0
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
requests==2.32.5
rich==14.2.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
